Refactor question server types
