+++ apps/prairielearn/src/schemas/infoElementCore.ts
@@ -37,28 +37,29 @@ const DependencyJsonSchema = z
   .strict()
   .describe("The element's client-side dependencies.");
 
+const DynamicDependencyJsonSchema = z
+  .object({
+    comment: CommentJsonSchema.optional(),
+    nodeModulesScripts: z
+      .record(z.string())
+      .describe('The scripts required by this element from /node_modules as an importmap.')
+      .optional(),
+    elementScripts: z
+      .record(z.string())
+      .describe(
+        "The scripts required by this element from the element's directory as an importmap.",
+      )
+      .optional(),
+  })
+  .strict()
+  .describe("The element's client-side dynamic dependencies.");
+
 export const ElementCoreJsonSchema = z
   .object({
     comment: CommentJsonSchema.optional(),
     controller: z.string().describe("The name of the element's controller file."),
     dependencies: DependencyJsonSchema.optional(),
-    dynamicDependencies: z
-      .object({
-        comment: CommentJsonSchema.optional(),
-        nodeModulesScripts: z
-          .record(z.string())
-          .describe('The scripts required by this element from /node_modules as an importmap.')
-          .optional(),
-        elementScripts: z
-          .record(z.string())
-          .describe(
-            "The scripts required by this element from the element's directory as an importmap.",
-          )
-          .optional(),
-      })
-      .strict()
-      .describe("The element's client-side dynamic dependencies.")
-      .optional(),
+    dynamicDependencies: DynamicDependencyJsonSchema.optional(),
     additionalNames: z
       .array(z.string().describe('A name for this element to be used in question HTML files.'))
       .describe('Any additional names to give this element, i.e. for backwards compatibility.')
